+++ Docs/Whats_New.md
@@ -18,10 +18,16 @@ MoltenVK 1.2.3
 
 Released TBA
 
+- Add support for extensions:
+	- `VK_EXT_swapchain_maintenance1`
+	- `VK_EXT_surface_maintenance1`
 - Fix issue where extension `VK_KHR_fragment_shader_barycentric` 
   was sometimes incorrectly disabled due to a Metal driver bug.
-- Work around problems with using explicit LoD with arrayed depth images
-  on Apple Silicon.
+- Detect when size of surface has changed under the covers.
+- Change rounding of surface size provided by Metal from truncation to rounding-with-half-to-even.
+- Queue submissions retain wait semaphores until `MTLCommandBuffer` finishes.
+- Work around problems with using explicit LoD with arrayed depth images on Apple Silicon.
+- Update `VK_MVK_MOLTENVK_SPEC_VERSION` to version `37`.
 
 
 
@@ -47,6 +53,7 @@ Released 2023/01/23
 - Update to latest SPIRV-Cross:
   - MSL: Add support for writable images in iOS Tier2 argument buffers.
   - MSL: Fix potentially uninitialized warnings.
+
 
 
 MoltenVK 1.2.1